修复类型强制转换的问题。

+++ framework-amqp-rabbit/src/main/java/com/ai/southernquiet/amqp/rabbit/AmqpMessageRecover.java
@@ -29,7 +29,7 @@ public class AmqpMessageRecover extends RepublishMessageRecoverer {
             }
 
             long expiry = (long) value;
-            return expiry + expiry * properties.getMultiplier();
+            return expiry + expiry * (long) properties.getMultiplier();
         });
 
         if (expiration < properties.getExpiration().toMillis()) {
